Move AWS bootstrap authenticator and verifier to pkg/bootstrap/awsbootstrap
The STS-based node bootstrap protocol (authenticator, verifier and
shared token format) moves from upup/pkg/fi/cloudup/awsup to a leaf
package, following the pkibootstrap layout. This lets nodeup build the
AWS authenticator without importing the full awsup cloud
implementation and all AWS SDK service clients it constructs.
